Fix integer issue

+++ lns/common/preprocessing/scale_od.py
@@ -29,7 +29,7 @@ def _scale_od_ped(path: str) -> Dataset:  # noqa
             image_annotations = []
             with open(label_path, "r") as file:
                 for line in file:
-                    bounds = Bounds2D(*[int(s) for s in line.split()[4:8]])
+                    bounds = Bounds2D(*[float(s) for s in line.split()[4:8]])
                     image_annotations.append(Object2D(bounds, 0))
             annotations[image_path] = image_annotations
